x86-64 compilation failure fix
Hi Eric,

I just realized that my recent patches of kdump for i386 broke kexec-tools
compilation on x86-64 machine. Here is a patch attached to fix the problem.

Thanks
Vivek

o This patch fixes a x86_64 compilation failure issue. Recent patches
  changed the function signature of funciton setup_linux_system_parameters().
  Additional parameter kexec_flags is being passed now. This patch makes
  sure that this function is called with appropriate parameters.

o x86_64 code also compiles i386 elf32 vmlinux loader code. i386 elf loader
  code inturn uses crashdump code (crashdump-x86.c). Modified x86_64
  makefile to compile crashdump-x86.c as well.

o phdr->vaddr size type unsigened int. Did minor modificaition to make sure
  it compiles without warnings in x86_64 context.

+++ kexec/arch/i386/crashdump-x86.c
@@ -635,7 +635,7 @@ static int prepare_crash_memory_elf32_headers(struct kexec_info *info,
 		if (mend <= (MAXMEM - 1))
 			phdr->p_vaddr = mstart + PAGE_OFFSET;
 		else
-			phdr->p_vaddr = ULONG_MAX;
+			phdr->p_vaddr = UINT_MAX;
 		phdr->p_paddr = mstart;
 		phdr->p_filesz	= phdr->p_memsz	= mend - mstart + 1;
 		/* Do we need any alignment of segments? */
